notebook_check_scraper_gpt/requirements.txt
requests
fake-useragent
python-decouple
pillow
html-text
pandas
beautifulsoup4
lxml
openpyxl
azure-cosmos
numpy==1.26.4